hide quick start enable btc wallet button if there is a mnemonic key already
